Proper Fullscreen Init
Before, initializing the app in fullscreen didn't actually initialize it in fullscreen. It made a window and then it immediately called 'Resize'. This caused some confusing resoure use as resources that were planned to be deleted suddenly were required again (resizing screenspace buffers and textures for example. Thanks to this, now this is fixed.